updated refactoring to the correct files. updated README.md
