+++ evals/cases/autodetermination-decision-001.yaml
