+++ evals/cases/autodetermination-decision-002.yaml
