+++ evals/cases/ocr-ner-001-a.yaml
