+++ evals/cases/ocr-ner-001-b.yaml
