+++ evals/cases/ocr-ner-002-a.yaml
